moved walkDirForContracts and promisifiedReadFile to ./lib, which is where our common web3 code lives
