+++ artifacts/contracts/GreeterStrIntNoImp1.sol/GreeterStrIntNoImp1.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../../build-info/54633465feb7d95fb8f16e752a7d8315.json"
+}
