+++ simba_hardhat/src/tasks/exportcontract.ts
@@ -12,6 +12,10 @@ import { StatusCodeError } from 'request-promise/errors';
 import {
     writeAndReturnASTSourceAndCompiler,
 } from "./lib/api"
+import {
+    promisifiedReadFile,
+    walkDirForContracts,
+} from "./lib";
 
 interface Data {
     [key: string]: any;
@@ -60,34 +64,6 @@ interface Request {
 //     return files;
 // }
 
-export const walkDirForContracts = (dir: string, extension: string): Promise<string[]> =>
-    new Promise((resolve, reject) => {
-        fs.readdir(dir, {withFileTypes: true}, async (err, entries) => {
-            if (err) {
-                return reject(err);
-            }
-
-            let files: string[] = [];
-
-            for (const entry of entries) {
-                if (entry.isFile()) {
-                    const filePath = path.join(dir, entry.name);
-                    if (!extension || (extension && path.parse(filePath).ext === extension)) {
-                        files.push(filePath);
-                    }
-                } else if (entry.isDirectory()) {
-                    try {
-                        const subFiles = await walkDirForContracts(path.join(dir, entry.name), extension);
-                        files = files.concat(subFiles);
-                    } catch (e) {
-                        reject(e);
-                    }
-                }
-            }
-
-            resolve(files);
-        });
-    });
 
 // async function asyncReadFile(
 //     filePath: fs.PathLike,
@@ -101,17 +77,6 @@ export const walkDirForContracts = (dir: string, extension: string): Promise<str
 //         return err as any;
 //     }
 // }
-
-export const promisifiedReadFile = (filePath: fs.PathLike, options: { encoding?: null; flag?: string }): Promise<Buffer> =>
-    new Promise((resolve, reject) => {
-        fs.readFile(filePath, options, (err: NodeJS.ErrnoException | null, data: Buffer) => {
-            if (err) {
-                return reject(err);
-            }
-            return resolve(data);
-        });
-    });
-
 
 const exportContract = async (
     hre: HardhatRuntimeEnvironment,
